Test that CellInfo is provided with no-location NRI
Test that if there is a non-null CellIdentity included
in NetworkRegistrationInfo, that a location-sanitized
version of the CellIdentity is included when the NRI
is location-sanitized.

Note, removing enforcement that PLMN-ID and Network Name
are redacted. These fields are roughly equivalent to
TM#getNetworkOperator() and TM#getNetworkOperatorName()
which are not considered location sensitive. The reality is
that while there could be some incidental location here,
it's mostly at the country level, and because this information
is so critical to the functionining of many systems, it's
impractical to hide it behind location. Note that PLMN and
Network operator name are also available in network scan
even when the location toggle is off as additional precedent.
Hence, updating this enforcement to align with others.

Bug: 357159744
Test: atest NetworkRegistrationInfoTest#testSanitizeLocationInfo
Flag: EXEMPT bugfix
Change-Id: I1107f8f9f6673978fd5b36ede40c335724e48bce