[SMSoIP] Moved TP-MR to a SmsDispatchersController
TP-MR is not in sync when sending sms over CS and PS alternatively as
mMessageRef is non-static variable in base class (SMSDispatcher) of both
ImsSmsDispatcher and GsmSmsDispatcher. So it is independent for the
instances of both of the classes.

Moved TP-MR variable to SmsDispatchersController so that it can be
common for both class's instances per phoneId.

Doc: go/tpmr-fix
Bug: 374942620
Test: atest SmsDispatchersControllerTest, atest ImsSmsDispatcherTest, atest GmsSmsDispatcherTest, atest CdmaSmsDispatcherTest
Flag: EXEMPT bugfix
Change-Id: I00c2b724cbc630dbf93c3478b40ddceafc08c682