DSDA: Resolve missed call not received for 2nd MT.
When we have an existing ringing/dialing call and we attempt to receive
another incoming call on a different sim, ensure that we log the missed
call. Previously, this was failing in the IMS call tracker but now we're
explicitly handling this in Telecom so that we can log the call.

Bug: 394161212
Flag: com.android.server.telecom.flags.enable_call_sequencing
Test: Manual verification for RINGING + RINGING and DIALING + RINGING to
verify that 2nd call shows up in the call log and that we are properly
disconnecting the call in Telecom.
Test: atest CtsTelecomCujTestCases:CallSequencingBasicCallTest

Change-Id: I54d80e1dbb8893cadb3726418646a2fe1df9e74f