Ensure the same OnSubscriptionChangedListener is reused in PhoneConfigurationManager.
This CL ensures that the same OnSubscriptionChangedListener is reused in PhoneConfigurationManager#maybeEnableCellularDSDASupport. This is to prevent an issue where we see too many listeners being registered by Telephony since we are now reusing the same listener in PCM rather than creating a new instance of OnSubscriptionChangedListener and using that each time maybeEnableCellularDSDASupport is invoked.

Test: PhoneConfigurationManagerTest and SimultaneousCallingTrackerTest
Flag: EXEMPT minor bug fix
Bug: 386769454
Change-Id: Ic98f3a850f7105d49819ac930d04ff4921363998