Add exception handling for radio power off during emergency call
This change is to add an exception handling to prevent a phone crash
when the emergency call ends due to an internal variable initialization
issue when a radio power off occurs for some reasons during an emergency
call. The change is to initialize the member variables in case where
there is no emergency call in progress.

Bug: 394984282
Flag: EXEMPT bugfix
Test: atest EmergencyStateTrackerTest
Test: manual (verified exitEmergencyCallbackMode is called while
emergency call is in progress)

Change-Id: Icc35c67be7d4c80631607355d949aa0d8df44ac8