Do not switch phone type to CDMA
As CDMA is deprecated, force the phone to stay in GSM (A broader term for
3GPP RATs like UMTS, LTE, NR, etc..) mode.

This is just the partial change to unblock the issue that vendor RIL
reports CDMA when it's actually unknown. The phone type clean up will
be merged in other CLs.

Fix: 397326176
Fix: 399469738
Bug: 395645219
Test: Data browsing, making phone calls, and other basic telephony tests
Flag: com.android.internal.telephony.flags.phone_type_cleanup
Change-Id: Ifd45db11d390735fc35197904556d5e37ff36d85