Add sorting by ascending order for repeated atom fields
We have below 2 fields, which are object of grouping. but order is matter for repeated data when it is used for dimension factor.

CarrierRoamingSatelliteSession
    - repeated int32 supported_satellite_services
SatelliteEntitlement
    - repeated int32 entitlement_service_type

To reduce the number of cases, it is needed to make sure the list of service types is reported as ascending order every time.

Consulted with Metrics Council

Flag: EXEMPT bug fix
Bug: 373149631
Test: atest SatelliteStatsTest
Change-Id: I064d1b0a7374719145b3ac14273eef4fb6efb1a5