Check emergency messaging support before showing SOS button
The current check for isSatelliteConnectedViaCarrierWithinHysteresisTime() does not consider the carrier's capability of supporting emergency messaging. This means that if the user is connected to carrier satellite within the hysteresis time, but the carrier does not support emergency messaging, the user may still be shown an SOS button on dialer.

This commit adds a new check to the SatelliteSOSMessageRecommender that ensures that the carrier supports emergency messaging before allowing the user to see the SOS button.

Flag: EXEMPT bugfix
Bug: 394177376
Test: Manual system tests
SatelliteSOSMessageRecommenderTest SatelliteManagerTestOnMockService

Change-Id: Iffabdea9f71e53fad12f700eb2ceb77ede1a5449