Move Work Back to Individual Threads
As part of an effort to remove a variety of threads,
a few very sluggish operations have been identified that
would block a shared worker thread for multiple seconds,
in particular binding to an ImsService and doing a full
query of the LPA status.

As a temporary measure, move these operations back onto
bespoke worker threads as part of the ongoing evaluation
of how we can safely reduce the thread count in telephony.

In addition, there is a 5-way deadlock that occurs when
the ImsResolver is on the WorkerThread. While this is possible
anyway, it is much rarer without the blocking on this thread.
(ref b/398985346 and ref b/393048045).

Bug: 390244513
Test: atest FrameworksTelephonyTests
Flag: com.android.internal.telephony.flags.thread_shred
Change-Id: I84ef38269936968492c43c11c6e7b4077cba3228