Ensure immediate delivery of WFC notification action
The 'Do Not Ask Again' button on the Emergency Network Notification
(Wi-Fi Calling) was observed to be non-responsive immediately upon click. Log analysis revealed that while the `notification_action_clicked` event was registered, the associated broadcast `SILENCE_WIFI_CALLING_NOTIFICATION` was experiencing significant delivery delays by the Activity Manager,
sometimes up to 14 minutes.

This delay prevented the `NotificationActionReceiver` from processing the action in a timely manner, resulting in the notification not being
silenced or dismissed as expected by the user.

To resolve this, `Intent.FLAG_RECEIVER_FOREGROUND` has been added to the
`Intent` carried by the `PendingIntent` for the 'Do Not Ask Again'
action within the `EmergencyNetworkNotification`. This flag instructs the system to prioritize the delivery of this broadcast, ensuring the user's interaction is processed immediately.

This change restores the intended functionality of the notification action, allowing users to effectively silence the Emergency Network Notification and improving overall user experience.

Flag: com.android.internal.telephony.flags.immediately_process_do_not_show_again_broadcast
Fixes: 368690328
Test: manual - use adb command to trigger the notification
Change-Id: Ibd68f1fc6b74fcd25b4a7a2b53bdc1401dc18b97