Reduce the scope of synchronized() blocks
The following changes are made in this cl:
- Scope of synchronized() blocks is only reduced to be around
  pendingMaps
- @GuardedBy("mLock") for methods is removed: If a method has @GuardedBy("mLock"), then the calling function should
acquire the lock and then call the protected method. Instead, it would
be much cleaner to remove @GuardedBy("mLock") annotation and acquire the
lock inside the method so that the code in calling function is clean.

Bug: 409610090
Test: 418863274
Test: atest
Flag: com.android.internal.telephony.flags.satellite_improve_multi_thread_design
Change-Id: I52b1717f067d09b6fe35622f2acaee3959e12f21