Avoid setEmergencyMode for SMS when radio is/will be turned off
This change is to prevent an exception from being thrown for
setEmergencyMode when making an emergency call in airplane mode and
sending an emergency SMS while the radio is off or being turned off, and
to prevent errors during the emergency call procedure.
This is because if setEmergencyMode is called while the radio is turning
on or turning off, the modem may not handle it properly and may throw an
exception. If this happens, the state managed by APDS may not match the
modem, which may cause problems in the subsequent emergency call
procedure.

Bug: 409539285
Flag: EXEMPT bugfix
Test: atest EmergencyStateTrackerTest
Test: manual (verified that the emergency call proceeds normally when an
emergency SMS is sent by ELS while making an emergency call in APM)

Change-Id: I772dbb999f13803b68e6f64cd765097ddde9e916