Allow carrier override for default IMS APN usage.
To work around a modem testing issue on a Wear device, the IMS APN needs to be removed for a particular carrier.  Currently Telephony prevents this from happening, and this change makes it possible to disable the default IMS APN for a particular carrier config.

Bug: 412950733
Flag: EXEMPT flagged by carrier config
Change-Id: I70fd5294d1d6ca7bc5598b05fd74f779466e8018