Use the stored emergency registration result instead of an unknown one
In certain carrier networks that support ELS, an emergency SMS(ELS) may
be transmitted when an emergency call is initiated, depending on its
requirements. The problem arises when, in such a sequence where
emergency SMS precedes the emergency call, the emergency call
incorrectly retrieves an unknown emergency registration result.
This leads to the device initiating an unnecessary emergency network
scan and placing the emergency call in the limited service, even when it
could have proceeded with the emergency call on the currently camped
network based on the information acquired by the emergency SMS.

This change aims to modify the APDS so that when the emergency sequence
involves an emergency SMS followed by an emergency call, the emergency
call can utilize the emergency registration result obtained by the
emergency SMS.

Bug: 419814299
Flag: EXEMPT bugfix
Test: atest EmergencyStateTrackerTest
Test: manual (verified emergency call on a carrier network that supports
ELS)

Change-Id: If0181370a6e7f2d49b2b2255f516fafdcc744aec