Do not notify the modem of an auto rejected incoming call status
If the UE receives a MT IMS call while a CS call is ongoing, IMS
automatically rejects the IMS call and informs the Telephony to log it
in the missed call list. The modem does not need to be notified of this
rejected call, as it is already disconnected. Notifying the modem of the
call state could create a Call ID conflict, because both CS and IMS call
might have the same Call ID (e.g., Call ID 1). This modification avoids
that conflict by preventing the unnecessary IMS call state from being
sent to the modem.

Bug: 425501132
Test: Manual test
Test: atest ImsPhoneCallTrackerTest
Flag: EXEMPT bugfix
Change-Id: I9de1e97b0e57c1f76c4b44700566efe6c16299b3