Exit SMS emergency mode before radio power off
If a user turns airplane mode on while the UE is in emergency mode for
SMS, the modem can't exit emergency mode after the completion of the
SMS, as AP can't send the request to the modem because radio power is
off. Then, UE rejects all incoming calls because of the emergency state
after turning airplane mode off.
This CL is to make the modem to exit emergency mode before the radio
power is off when turning airplane mode on.

Bug: 429607770
Test: atest EmergencyStateTrackerTest
Test: manual test on O2 UK live network
Flag: EXEMPT bugfix
Change-Id: I9469a2bf1a8fea96ecd6d782e82014d9c8d2a35e