Always use TelephonyCapabilities to check for C/D/M support
Instead of running the minimalTelephonyCdmCheck() and then checking
for the PackageManager feature flags in different places, provide and
use generic methods to do it in TelephonyCapabilities.

This is only a code refactor, no logic has changed.

The FeatureFlags variable in minimalTelephonyCdmCheck() is at this
point unused and could be removed, but a followup change will use it
again, so leaving it on purpose.

Bug: 395087895
Test: mm && atest
Flag: EXEMPT minor refactor

Change-Id: I2840045a98d956e5ea0d98a249f23fb8a785cdfe