Skip Data Stall Recovery for Non-Active Sub
Skip data stall recovery actions for the non-active
sub. It's possible for a DSDA device to have data;
however, for now we can assume DSDS devices, and that
data connectivity on the non-default sub will fail
for a variety of reasons. In this case, it fails due
to an active voice call; however, by the time the
voice call status is checked it's already doing
data stall recovery.

Bug: 430976119
Test: atest DataStallRecoveryManagerTest
Flag: com.android.internal.telephony.flags.inactive_data_network_is_not_stalled
Change-Id: I70e7a54ba94de4ba842d62ddef0056b0cff8a7c2