Fix to not mark MT CS call as an emergency call
APDS updates an emergency call information if the call is not recognized
as an emergency call to handle SIP 380 response. The information is used
in RIL/CP IMS when a normal routing emergency call is redailed over
emergency routing emergency call. However, the information is updated
accidently when UE receives CS MT call. This CL is to fix the problem by
updating emergency information only for an emergency routing emergency
calls.

Bug: 443326734
Test: atest ConnectionTest
Test: manual test on test platform(b/443647857)
Flag: EXEMPT bugfix
Change-Id: I01e505e97870f72a68849e58b2096ea212b919dd